refactor: move functions around to better group actions
